Make generic output handle string and directory
