+++ src/renderer/components/node/NodeOutputs.tsx
@@ -8,12 +8,10 @@ import { assertNever } from '../../../common/util';
 import { ExecutionContext } from '../../contexts/ExecutionContext';
 import { GlobalContext } from '../../contexts/GlobalNodeState';
 import { DefaultImageOutput } from '../outputs/DefaultImageOutput';
-import { DirectoryOutput } from '../outputs/DirectoryOutput';
 import { GenericOutput } from '../outputs/GenericOutput';
 import { LargeImageOutput } from '../outputs/LargeImageOutput';
 import { OutputContainer } from '../outputs/OutputContainer';
 import { OutputProps } from '../outputs/props';
-import { TextOutput } from '../outputs/TextOutput';
 
 interface FullOutputProps extends Omit<Output, 'id' | 'type'>, OutputProps {
     definitionType: Type;
@@ -33,10 +31,10 @@ const pickOutput = (kind: OutputKind, props: FullOutputProps) => {
             isGenericType = false;
             break;
         case 'directory':
-            OutputType = DirectoryOutput;
+            OutputType = GenericOutput;
             break;
         case 'text':
-            OutputType = TextOutput;
+            OutputType = GenericOutput;
             break;
         case 'generic':
             OutputType = GenericOutput;
